requirements.txt
fastapi
uvicorn
pandas
scikit-learn
transformers
torch
groq
python-dotenv
streamlit
requests
pytest
